implement missed PeerConnection.Observer onConnectionChange
`onConnectionChange` is the `Java Interface Default Method`
introduced in Java 8, which aims to add new method to interface
and keep backward compatibility. In theory, we should not have to
implement this method.

Seems the following section in app/build.gradle does not work.

```
compileOptions {
    sourceCompatibility JavaVersion.VERSION_1_8
    targetCompatibility JavaVersion.VERSION_1_8
}
```

Weird, not sure why, just happened when upgrade to RN 0.63.